feat(updateUser): userResoler, authorized mutations

+++ src/resolvers/UserResolver.ts
@@ -1,7 +1,7 @@
-import { ApolloError } from 'apollo-server-express';
+import { ApolloError, AuthenticationError, ForbiddenError } from 'apollo-server-express';
 
 import { Knex } from 'knex';
-import { Query, Resolver, Ctx, Arg, Mutation, ObjectType, Field, Authorized } from 'type-graphql';
+import { Query, Resolver, Ctx, Arg, Mutation, ObjectType, Field, Authorized, FieldResolver, Root } from 'type-graphql';
 import { LoginInput } from '../dto/LoginInput';
 import { RegisterInput } from '../dto/RegisterInput';
 import { User } from '../entities/User';
@@ -10,6 +10,11 @@ import * as jwt from 'jsonwebtoken';
 import { SECRET } from '../constants/constants';
 import { ChangePasswordInput } from '../dto/ChangePassword';
 import { UpdateInput } from '../dto/UpdateInput';
+import { ArgUserInfo } from '../dto/ArgUserInfo';
+
+import { Post } from '../entities/Post';
+import { MyContext } from '../@types/types';
+
 @ObjectType()
 class AuthResponse {
   @Field()
@@ -19,13 +24,25 @@ class AuthResponse {
   user: User;
 }
 
-@Resolver()
+@Resolver(() => User)
 export class UserResolver {
   // constructor(private bookService: BookService) {}
+
+  @FieldResolver(() => [Post])
+  async posts(@Root() user: User, @Ctx() ctx: MyContext) {
+    const {
+      dataLoader: { postLoader },
+    } = ctx;
+    const posts = await postLoader.load(user.user_id);
+
+    return posts;
+  }
+
   @Query(() => [User])
-  async users(@Ctx() ctx) {
+  async users(@Ctx() ctx, @Arg('info') info: ArgUserInfo) {
     const db: Knex = ctx.db;
-    const users = await db.select().table('users');
+    const { limit, offset, email } = info;
+    const users = await db.select().table('users').where('email', 'like', `%${email}%`).limit(limit).offset(offset);
     console.log({ users });
     return users;
   }
@@ -34,13 +51,13 @@ export class UserResolver {
   @Authorized()
   async changePassword(@Arg('changePassInput') changePassInput: ChangePasswordInput, @Ctx() ctx) {
     const db: Knex = ctx.db;
-    const { user_id, currentPass, newPass, confirmPass } = changePassInput;
+    const { username, currentPass, newPass, confirmPass } = changePassInput;
     try {
       if (newPass !== confirmPass) {
         throw new ApolloError('Password and confirm does not match');
       }
 
-      const [matchUser] = await db('users').where({ user_id: user_id });
+      const [matchUser] = await db('users').where({ username: username });
       if (!matchUser) {
         throw new ApolloError('User not found');
       }
@@ -52,7 +69,7 @@ export class UserResolver {
       const hashPass = await bcrypt.hash(newPass, 10);
 
       const [user] = await db('users')
-        .where({ user_id: user_id })
+        .where({ username: username })
         .update({
           password: hashPass,
         })
@@ -71,11 +88,11 @@ export class UserResolver {
     try {
       const [user] = await db('users').where({ username });
       if (!user) {
-        throw new ApolloError('User not found');
+        throw new AuthenticationError('User not found');
       }
       const matchPass = await bcrypt.compare(password, user.password);
       if (!matchPass) {
-        throw new ApolloError('Username/ password wrong');
+        throw new AuthenticationError('Username/ password wrong');
       }
 
       const token = jwt.sign({ id: user.user_id }, SECRET, {
@@ -83,7 +100,7 @@ export class UserResolver {
       });
       return { token, user };
     } catch (error) {
-      throw new ApolloError(error.message);
+      throw new AuthenticationError(error.message);
     }
   }
   @Mutation(() => User)
@@ -104,6 +121,8 @@ export class UserResolver {
     console.log({ newUser });
     return newUser;
   }
+
+  /*
   @Mutation(() => User)
   async addUser(@Arg('input') input: RegisterInput, @Ctx() ctx) {
     const db: Knex = ctx.db;
@@ -111,16 +130,21 @@ export class UserResolver {
     console.log(user);
     return user;
   }
+  */
+
   @Mutation(() => User)
-  async updateUser(@Arg('userId') userId: string, @Arg('payload') payload: UpdateInput, @Ctx() ctx) {
-    const db: Knex = ctx.db;
-    const { user_id, ...userInfo } = payload;
+  @Authorized()
+  async updateUser(@Arg('userId') userId: string, @Arg('payload') payload: UpdateInput, @Ctx() ctx: MyContext) {
+    const { db, userId: userIdCtx } = ctx;
+    const { ...userInfo } = payload;
     try {
       const [matchUser] = await db('users').where({ user_id: userId });
       if (!matchUser) {
         throw new ApolloError('User not found');
       }
-
+      if (matchUser.user_id != userIdCtx) {
+        throw new ForbiddenError('Permission required');
+      }
       const timestamp = Date.now();
 
       const [user] = await db('users')
@@ -134,12 +158,16 @@ export class UserResolver {
     }
   }
   @Mutation(() => User)
-  async deleteUser(@Arg('userId') userId: string, @Ctx() ctx) {
-    const db: Knex = ctx.db;
+  @Authorized()
+  async deleteUser(@Arg('userId') userId: string, @Ctx() ctx: MyContext) {
+    const { db, userId: userIdCtx } = ctx;
     try {
       const [matchUser] = await db('users').where({ user_id: userId });
       if (!matchUser) {
         throw new ApolloError('User not found');
+      }
+      if (matchUser.user_id != userIdCtx) {
+        throw new ForbiddenError('Permission required');
       }
       const [user] = await db('users').where({ user_id: userId }).delete().returning('*');
       console.log('deleted', user);
